+++ src/main/java/eu/yals/exception/IdentNotFoundException.java
@@ -1,0 +1,5 @@
+package eu.yals.exception;
+
+//TODO javadoc
+public class IdentNotFoundException extends YalsException {
+}
