+++ src/main/java/eu/yals/exception/PageNotFoundException.java
@@ -1,0 +1,5 @@
+package eu.yals.exception;
+
+//TODO javadoc
+public class PageNotFoundException extends YalsException {
+}
