+++ src/test/java/eu/yals/test/pageobjects/NotFoundViewPageObject.java
@@ -1,12 +1,12 @@
 package eu.yals.test.pageobjects;
 
 import com.codeborne.selenide.SelenideElement;
-import eu.yals.ui.err.NotFoundView;
+import eu.yals.ui.err.IdentNotFoundView;
 
 import static com.codeborne.selenide.Selenide.$;
 
 /**
- * Page Object for {@link NotFoundView}. Contains elements from NotFoundView
+ * Page Object for {@link IdentNotFoundView}. Contains elements from NotFoundView
  *
  * @since 2.7
  */
